+++ fjs/cas/module.f.ts
@@ -7,7 +7,7 @@ import { sha256, type Sha2, type State as Sha2State } from '../crypto/sha2/modul
 import { join, normalize, parse } from '../path/module.f.ts'
 import { empty, length, maxLength, maxLengthBytes, msb, vec, type Vec } from '../types/bit_vec/module.f.ts'
 import { cBase32ToVec, vecToCBase32 } from '../basen/cbase32/module.f.ts'
-import { foldStep, forEachStep, frameStep, okStep, pure, step, type Effect, type Operation } from '../effects/module.f.ts'
+import { foldStep, forEachStep, history, historyStep, okStep, pure, step, type Effect, type Operation } from '../effects/module.f.ts'
 import { eff } from '../effects/eff/module.f.ts'
 import {
     access,
@@ -242,12 +242,12 @@ export const fileCas = (sha2: Sha2) => (path: string): FileCas => {
                                             const nextPath = step(
                                                 now(),
                                                 t => pure(join(stageDir, stageName(t + leaseDelta, rndStr))))
-                                            const renamed = frameStep(
-                                                nextPath,
+                                            const renamed = historyStep(
+                                                history(nextPath),
                                                 next => rename(curPath, next))
                                             return step(
                                                 renamed,
-                                                ({ result: [rt, v], param: next }) =>
+                                                ([[rt, v], next]) =>
                                                     rt === 'error'
                                                         ? fail(curPath, v)
                                                         : loop(newState, newOffset, next)(tail))